+++ src/components/pages/RecruitingDetail/detail/index.js
@@ -1,4 +1,3 @@
 export { default as ApplicantListSection } from './ApplicantListSection';
 export { default as RecruitingHeader } from './RecruitingHeader';
 export { default as RecruitingInfoSection } from './RecruitingInfoSection';
-export { default as ApplicantCard } from './ApplicantCard';
