--- src/components/pages/RecruitingDetail/filters/index.js
@@ -1,1 +1,0 @@
-export { default as ApplicantFilters } from './ApplicantFilters';
