+++ src/components/pages/RecruitingDetail/index.js
@@ -3,12 +3,3 @@ export * from './detail';
 
 // Modal components  
 export * from './modals';
-
-// Stats components
-export * from './stats';
-
-// Evaluation components
-export * from './evaluation';
-
-// Filter components
-export * from './filters';
