+++ src/components/pages/RecruitingDetail/modals/index.js
@@ -1,3 +1,2 @@
 export { default as BulkStatusChangeModal } from './BulkStatusChangeModal';
-export { default as ApplicantModal } from './ApplicantModal';
 export { default as EmailModal } from './EmailModal';
