+++ src/components/recruiting/applicant/index.js
@@ -1,0 +1,3 @@
+export { default as ApplicantCard } from './ApplicantCard';
+export { default as ApplicantFilters } from './ApplicantFilters'; 
+export { default as ApplicantModal } from './ApplicantModal';
