+++ src/components/recruiting/index.js
@@ -1,0 +1,4 @@
+// Recruiting Domain Components
+export * from './applicant';
+export * from './evaluation'; 
+export * from './stats';
