+++ src/pages/RecruitingDetailPage.jsx
@@ -3,8 +3,9 @@ import { useParams, useNavigate } from 'react-router-dom';
 import { RecruitingDetailProvider, useRecruitingDetail } from '../contexts/RecruitingDetailContext.jsx';
 import { EvaluationProvider, useEvaluation } from '../contexts/EvaluationContext.jsx';
 import { AdminHeader } from '../components/admin';
-import { StatsSection } from '../components/pages/RecruitingDetail/stats';
-import { ApplicantModal, EmailModal, BulkStatusChangeModal } from '../components/pages/RecruitingDetail/modals';
+import { StatsSection } from '../components/recruiting';
+import { ApplicantModal } from '../components/recruiting';
+import { EmailModal, BulkStatusChangeModal } from '../components/pages/RecruitingDetail/modals';
 import DeleteConfirmModal from '../components/common/DeleteConfirmModal';
 import { RecruitingInfoSection, RecruitingHeader, ApplicantListSection } from '../components/pages/RecruitingDetail/detail';
 import { RECRUITMENT_CONFIG, SORT_OPTIONS } from '../constants/recruitment';
